refactoring -> new project structure
