+++ flinkfood-demo/CustomerViewJob/dependency-reduced-pom.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.flinkfood.jobs.restaurant</groupId>
-  <artifactId>customerview</artifactId>
+  <groupId>org.flinkfood.jobs.customer</groupId>
+  <artifactId>CustomerView</artifactId>
   <name>customer-view</name>
   <version>1.0</version>
   <build>
