+++ flinkfood-demo/flinkfood-v-DataStream/src/main/java/org/flinkfood/flinkjobs/AggregateFunction.java
@@ -1,0 +1,11 @@
+package org.flinkfood.flinkjobs;
+
+import org.apache.flink.shaded.zookeeper3.org.apache.zookeeper.Transaction;
+import org.apache.flink.streaming.api.functions.KeyedProcessFunction;
+import org.apache.flink.util.Collector;
+import org.flinkfood.schemas.Address;
+import org.flinkfood.schemas.Customer;
+
+public class AggregateFunction {
+
+}
