+++ flinkfood-demo/flinkfood-v-DataStream/src/main/java/org/flinkfood/schemas/Customer.java
@@ -1,0 +1,50 @@
+package org.flinkfood.schemas;
+
+public class Customer {
+
+    public int id;
+    public String name;
+    public Character gender;
+    public int age;
+
+    public Customer() {
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Character getGender() {
+        return gender;
+    }
+
+    public void setGender(Character gender) {
+        this.gender = gender;
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    @Override
+    public String toString() {
+        return "Customer [id=" + id + ", name=" + name + ", gender=" + gender + ", age=" + age + "]";
+    }
+
+}
