--- flinkfood-demo/src/main/java/org/flinkfood/flinkjobs/Customer.java
@@ -1,17 +1,0 @@
-package org.flinkfood.flinkjobs;
-import java.util.Date;
-
-public class Customer {
-
-    public int id;
-    public String username;
-    public String first_name;
-    public String last_name;
-    public Date birthdate;
-    public String email;
-    public String fiscal_code;
-
-    public Customer() {
-    }
-    
-}
