--- flinkfood-demo/src/main/java/org/flinkfood/flinkjobs/Customer_address.java
@@ -1,16 +1,0 @@
-package org.flinkfood.flinkjobs;
-
-public class Customer_address {
-
-    public int id;
-    public int customer_id;
-    public String street;
-    public String address_number;
-    public int zip_code;
-    public String city;
-    public String province;
-    public String country;
-
-    public Customer_address() {
-    }
-}
