--- flinkfood-demo/src/main/java/org/flinkfood/schemas/Dish.java
@@ -1,71 +1,0 @@
-package org.flinkfood.schemas;
-
-public class Dish {
-
-    public int id;
-    public int restaurant_id;
-    public String name;
-    public String price;
-    public String currency;
-    public String category;
-    public String description;
-
-    public Dish() {
-    }
-
-    public int getId() {
-        return this.id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public int getRestaurant_id() {
-        return this.restaurant_id;
-    }
-
-    public void setRestaurant_id(int restaurant_id) {
-        this.restaurant_id = restaurant_id;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getPrice() {
-        return this.price;
-    }
-
-    public void setPrice(String price) {
-        this.price = price;
-    }
-
-    public String getCurrency() {
-        return this.currency;
-    }
-
-    public void setCurrency(String currency) {
-        this.currency = currency;
-    }
-
-    public String getCategory() {
-        return this.category;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-
-    public String getDescription() {
-        return this.description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-}
